refactor: remove useAnalytics shim, use CHART_TEXT token in chartUtils
- Delete frontend/src/hooks/useAnalytics.ts (backwards-compat re-export
  shim flagged by our own "don't keep shims" rule); rewrite 16 callers
  to import directly from @/hooks/api/useAnalytics.
- chartUtils.barLabelStyle.fill: swap the hardcoded "#f5f5f7" for the
  CHART_TEXT.secondary design token so colors stay in sync.
