+++ frontend/src/components/analytics/CashFlowForecast.tsx
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion'
 import { TrendingUp, TrendingDown, AlertTriangle } from 'lucide-react'
 import { useMemo } from 'react'
-import { useMonthlyAggregation } from '@/hooks/useAnalytics'
+import { useMonthlyAggregation } from '@/hooks/api/useAnalytics'
 import { AreaChart, Area, XAxis, YAxis, CartesianGrid, Tooltip, ReferenceLine } from 'recharts'
 import { formatCurrency, formatCurrencyShort } from '@/lib/formatters'
 import { chartTooltipProps, ChartContainer } from '@/components/ui'
