+++ frontend/src/components/analytics/CreditCardHealth.tsx
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion'
 import { CreditCard, AlertTriangle, CheckCircle } from 'lucide-react'
 import { useMemo } from 'react'
-import { useAccountBalances } from '@/hooks/useAnalytics'
+import { useAccountBalances } from '@/hooks/api/useAnalytics'
 import { formatCurrency, formatPercent } from '@/lib/formatters'
 import { usePreferencesStore, selectCreditCardLimits } from '@/store/preferencesStore'
