+++ frontend/src/components/analytics/PeriodComparison.tsx
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion'
 import { TrendingUp, TrendingDown, Minus, Zap, Calendar, ArrowLeftRight } from 'lucide-react'
 import { useState, useMemo, useCallback } from 'react'
-import { useMonthlyAggregation } from '@/hooks/useAnalytics'
+import { useMonthlyAggregation } from '@/hooks/api/useAnalytics'
 import { useTransactions } from '@/hooks/api/useTransactions'
 import { formatCurrency } from '@/lib/formatters'
 import { rawColors } from '@/constants/colors'
