+++ frontend/src/components/upload/AccountClassifier.tsx
@@ -1,6 +1,6 @@
 import { motion } from 'framer-motion'
 import { Check, Info, Settings2 } from 'lucide-react'
-import { useAccountBalances } from '@/hooks/useAnalytics'
+import { useAccountBalances } from '@/hooks/api/useAnalytics'
 import { useAccountStore } from '@/store/accountStore'
 import type { AccountType } from '@/types'
 import { toast } from 'sonner'
