--- frontend/src/hooks/useAnalytics.ts
@@ -1,19 +1,0 @@
-/**
- * Re-export all analytics hooks from the consolidated location
- * This file is kept for backwards compatibility
- */
-export {
-  useKPIs,
-  useRecentTransactions,
-  useOverview,
-  useBehavior,
-  useTrends,
-  useAccountDistribution,
-  useCategoriesChart,
-  useMonthlyTrends,
-  useCategoryBreakdown,
-  useAccountBalances,
-  useMonthlyAggregation,
-  useTotals,
-  useMasterCategories,
-} from './api/useAnalytics'
