+++ frontend/src/pages/InvestmentAnalyticsPage.tsx
@@ -3,7 +3,7 @@ import { rawColors } from '@/constants/colors'
 import { TrendingUp, PieChart, DollarSign, LineChart, Target } from 'lucide-react'
 import MetricCard from '@/components/shared/MetricCard'
 import { useMemo, useState } from 'react'
-import { useAccountBalances } from '@/hooks/useAnalytics'
+import { useAccountBalances } from '@/hooks/api/useAnalytics'
 import { useTransactions } from '@/hooks/api/useTransactions'
 import { usePreferences } from '@/hooks/api/usePreferences'
 import { PieChart as RechartsPie, Pie, Cell, Tooltip, Legend, AreaChart, Area, XAxis, YAxis, CartesianGrid } from 'recharts'
