+++ frontend/src/pages/NetWorthPage.tsx
@@ -2,7 +2,7 @@ import { motion } from 'framer-motion'
 import { rawColors } from '@/constants/colors'
 import { TrendingUp, PiggyBank, CreditCard, BarChart3, ChevronDown, ChevronRight, type LucideIcon } from 'lucide-react'
 import MetricCard from '@/components/shared/MetricCard'
-import { useAccountBalances } from '@/hooks/useAnalytics'
+import { useAccountBalances } from '@/hooks/api/useAnalytics'
 import { useChartDimensions } from '@/hooks/useChartDimensions'
 import { useTransactions } from '@/hooks/api/useTransactions'
 import { usePreferences } from '@/hooks/api/usePreferences'
